36 make new layouts pages

+++ Calendary.Ng/src/app/components/calendar-images/calendar-images.component.ts
@@ -1,43 +1,46 @@
 import { CommonModule } from '@angular/common';
-import { Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
+import { Component, ElementRef, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
 import { FileUploader, FileUploadModule } from 'ng2-file-upload';
 import { ImageService } from '../../../services/image.service';
 import { Image } from '../../../models/image';
 import { MatIconModule } from '@angular/material/icon';
 import { CalendarService } from '../../../services/calendar.service';
+import { ImageItemComponent } from '../image-item/image-item.component';
 @Component({
   selector: 'app-calendar-images',
   standalone: true,
-  imports: [CommonModule, FileUploadModule, MatIconModule],
+  imports: [CommonModule, FileUploadModule, MatIconModule, ImageItemComponent],
   templateUrl: './calendar-images.component.html',
   styleUrl: './calendar-images.component.scss'
 })
 export class CalendarImagesComponent implements OnChanges, OnInit {
-  @Input() calendarId!: number;
 
+
+  @Input() calendarId!: number;
+  
+  @Output() imagesFull = new EventEmitter<boolean>();
+
+  @ViewChild('fileInput') fileInput!: ElementRef;
+
+  items = Array.from({length: 12}, (_, i) => i + 1);
+  monthes = ['Січень', 'Лютий', 'Березень', 'Квітень', 'Травень', 'Червень', 'Липень', 'Серпень', 'Вересень', 'Жовтень', 'Листопад', 'Грудень'];
   images: Image[] = [];
 
   public uploader!: FileUploader;
+  selectedMonth: number = 0;
   
-  @Output() uploadCompleted = new EventEmitter<string>();
-  @Output() generatedCompleted = new EventEmitter<boolean>();
-
-
-  constructor(private imageService : ImageService, 
-    private calendarService: CalendarService
-  ) 
-  {
-   
-  }
+  constructor(
+    private imageService : ImageService) 
+  { }
 
   ngOnInit(): void {
     this.initializeUploader();
     this.loadImages();
   }
 
-
   ngOnChanges(changes: SimpleChanges): void {
-    if (changes['calendarId'] && !changes['calendarId'].firstChange) {
+    if (changes['calendarId'] && !changes['calendarId'].firstChange) 
+    {
       this.initializeUploader();
     }
   }
@@ -52,6 +55,10 @@ export class CalendarImagesComponent implements OnChanges, OnInit {
         maxFileSize: 5 * 1024 * 1024 // Максимальний розмір 5MB
       });
 
+      this.uploader.onBeforeUploadItem = (item) => {
+        item.url = `/api/image/upload/${this.calendarId}?month=${this.selectedMonth}`;
+      };
+
       this.uploader.onCompleteItem = (item, response, status, headers) => {
         this.loadImages();
       };
@@ -62,11 +69,16 @@ export class CalendarImagesComponent implements OnChanges, OnInit {
     }
   }
 
+  getImage(monthNumber : number): Image | undefined {
+    return this.images.find(i => i.monthNumber === monthNumber);
+  }
+
   loadImages(): void {
     // Assuming imageService is injected and has a method to get images by calendarId
     this.imageService.getImages(this.calendarId).subscribe(
       (images: Image[]) => {
         this.images = images;
+        this.imagesFull.emit(this.images.length === 12);
       },
       (error) => {
         console.error('Error loading images:', error);
@@ -74,8 +86,17 @@ export class CalendarImagesComponent implements OnChanges, OnInit {
     );
   }
 
-  deleteImage(index: number): void {
-    const image = this.images[index];
+  onStartUpload(monthNumber: number): void 
+  {
+    this.selectedMonth = monthNumber;
+    this.fileInput.nativeElement.click();
+  }
+
+  deleteImage(monthNumber: number): void {
+    const image = this.images.find(i => i.monthNumber === monthNumber);
+    if (!image) {
+      return;
+    }
     this.imageService.deleteImage(image.id).subscribe(() => {
       this.loadImages();
     });
@@ -83,12 +104,5 @@ export class CalendarImagesComponent implements OnChanges, OnInit {
 
   canGeneratePdf(): boolean {
     return this.images.length === 12;
-  }
-
-  generatePdf(): void {
-    // Виклик API для генерації PDF
-    this.calendarService.generatePdf(this.calendarId).subscribe(response => {
-      this.generatedCompleted.emit(true);
-    });
   }
 }
